+++ wayland-client/src/rust_imp/proxy.rs
@@ -3,6 +3,7 @@ use std::sync::{Arc, Mutex};
 
 use wayland_commons::map::{Object, ObjectMap, ObjectMetadata};
 use wayland_commons::utils::UserData;
+use wayland_commons::wire::{Argument, ArgumentType};
 use wayland_commons::MessageGroup;
 
 use super::connection::Connection;
@@ -143,6 +144,84 @@ impl ProxyInner {
         }
     }
 
+    pub(crate) fn send_constructor<I, J>(
+        &self,
+        msg: I::Request,
+        version: Option<u32>,
+    ) -> Result<NewProxyInner, ()>
+    where
+        I: Interface,
+        J: Interface,
+    {
+        // grab the connection lock before anything else
+        // this avoids the risk or races during object creation
+        let mut conn_lock = self.connection.lock().unwrap();
+        if !self.is_alive() {
+            return Err(());
+        }
+        let destructor = msg.is_destructor();
+        let mut msg = msg.into_raw(self.id);
+        if ::std::env::var_os("WAYLAND_DEBUG").is_some() {
+            println!(
+                " -> {}@{}: {} {:?}",
+                I::NAME,
+                self.id,
+                self.object.requests[msg.opcode as usize].name,
+                msg.args
+            );
+        }
+
+        let opcode = msg.opcode;
+
+        // sanity check
+        let mut nid_idx = I::Request::MESSAGES[opcode as usize]
+            .signature
+            .iter()
+            .position(|&t| t == ArgumentType::NewId)
+            .expect("Trying to use 'send_constructor' with a message not creating any object.");
+
+        if let Some(o) = I::Request::child(opcode, 1, &()) {
+            if !o.is_interface::<J>() {
+                panic!("Trying to use 'send_constructor' with the wrong return type. Required interface {} but the message creates interface {}")
+            }
+        } else {
+            // there is no target interface in the protocol, this is a generic object-creating
+            // function (likely wl_registry.bind), the newid arg will thus expand to (str, u32, obj)
+            nid_idx += 2;
+        }
+        // insert the newly created object in the message
+        let newproxy = match &mut msg.args[nid_idx] {
+            &mut Argument::NewId(ref mut newid) => {
+                let newp = match version {
+                    Some(v) => self.child_versioned::<J>(v),
+                    None => self.child::<J>(),
+                };
+                *newid = newp.id;
+                newp
+            }
+            _ => unreachable!(),
+        };
+
+        let _ = conn_lock.write_message(&msg).expect("Sending a message failed.");
+        if destructor {
+            self.object.meta.alive.store(false, Ordering::Release);
+            {
+                // cleanup the map as appropriate
+                let mut map = conn_lock.map.lock().unwrap();
+                let server_destroyed = map
+                    .with(self.id, |obj| {
+                        obj.meta.client_destroyed = true;
+                        obj.meta.server_destroyed
+                    }).unwrap_or(false);
+                if server_destroyed {
+                    map.remove(self.id);
+                }
+            }
+        }
+
+        Ok(newproxy)
+    }
+
     pub(crate) fn equals(&self, other: &ProxyInner) -> bool {
         self.is_alive() && Arc::ptr_eq(&self.object.meta.alive, &other.object.meta.alive)
     }
@@ -157,13 +236,22 @@ impl ProxyInner {
         self.child_versioned::<I>(self.object.version)
     }
 
-    pub fn child_versioned<I: Interface>(&self, version: u32) -> NewProxyInner {
+    pub(crate) fn child_versioned<I: Interface>(&self, version: u32) -> NewProxyInner {
         let new_object = Object::from_interface::<I>(version, self.object.meta.child());
         let new_id = self.map.lock().unwrap().client_insert_new(new_object);
         NewProxyInner {
             map: self.map.clone(),
             connection: self.connection.clone(),
             id: new_id,
+        }
+    }
+
+    pub(crate) fn child_placeholder(&self) -> ProxyInner {
+        ProxyInner {
+            map: self.map.clone(),
+            connection: self.connection.clone(),
+            object: Object::placeholder(self.object.meta.child()),
+            id: 0,
         }
     }
 }